Remove remaining unused model files
