+++ client/src/components/History/model/index.js
@@ -1,12 +1,6 @@
-export { Content } from "./Content";
-export { Dataset } from "./Dataset";
-export { DatasetCollection } from "./DatasetCollection";
-export { JobStateSummary } from "./JobStateSummary";
-export { DateStore } from "./DateStore";
 export { History } from "./History";
-export { STATES } from "./states";
 
-// crud functions, ajax call + cache
+// crud functions post bulk requests
 export {
     // operations on lists
     hideSelectedContent,
